fixing release to branch
Unfortunately we could completely use GOV.UK Frontend code without
a small modification.

The issue was that GOV.UK Frontend versions everthing inside their `package/`
folder. This meant that the release-to-branch.sh was pushing up package/govuk-frontend
to githube. However in our case we are not interested in versioning our compiled
code because the compiled code is only used when we publish a release.

So to get this to work we force add `package/` just for the new pre-release
branch we use for testing.