Adds script to tag any merges to master
For use by a Jenkins job (similar to the FE toolkit auto-tag job)
which polls for scm changes every 2 minutes.

The tag version number is taken from the packaged release, as
the top level package.json does not have a version number. This
should be fine in practice as the Jenkins polling job can catch/ignore
failures if the tag already exists.

The tag message includes a link to the Changelog rather than
copying across the full details.