Copy govuk-frontend to src folder
When creating a new component, I found that it would be difficult to
get it to work for testing purposes with relative paths because the
GOV.UK Frontend components would be in a completely different path between
developing our own component and once we deploy our components. To simplify
the process I decided to try mirror the package directory during local
dev and test.