Create Digital Marketplace Footer component
This component has created because we wanted a centralised footer component
which would standardise our footer content and links. This is
what we used to have with digitalmarketplace-frontend-toolkit but the
difference is that we are no longer adding our own footer styling/markup
it all uses GOV.UK Frontend Footer component
https://design-system.service.gov.uk/components/footer/

To use, import the macro an then call dmFooter({}) whereever you need
the footer component to be used.