Refactor the component
To simplify the logic, we decided it would be better to pass in a role.
If the user is logged in they will always be given a role. Therefore,
we do not need to know whether they have logged in we can work that out
by checking if the role has been set.

This commit also means that we do not have to set the context when we
import this component and do not have to worry about `current_user` object

Eg.

{{ dmHeader({
  "role": current_user.role | default(None)
})}}

or

{{ dmHeader({
  "isAdmin: "true"
  "role": current_user.role | default(None)
})}}