Add puppeteer to enable/disable headless mode for chrome
By default puppeteer runs chrome in headless mode. The
developer can choose whether or not they would like to see the browser
while the tests are running. This can be useful when debugging.

To turn off headless mode:
create/set an environment variable called HEADLESS to any value other
then false. Run the tests and the browser should appear and be visible.