Add cookie settings full page example
Adds an example of how the cookie settings JavaScript should be used,
complete with alerts, buttons, and radios oh my!

There's also a sneaky link in the footer for clearing the cookie
settings to be able to test it multiple times without fiddling with
browser settings.

The page layout is copied from [digitalmarketplace-user-frontend][1].

[1]: https://github.com/alphagov/digitalmarketplace-user-frontend/blob/release-260/app/templates/cookies/cookie_settings.html