Ignore no-var linting rule
This fixes the warnings (for standard 17+, errors) we get because of our use of "var" instead of
let/const.

standard doesn't offer an obvious way to ignore a specific rule in the config, and I'd prefer
not to go down the route of defining a custom eslint config since that kind of defeats the purpose
of standard.

Ignoring its rules also kind of defeats that purpose, but at least we're aware of this one specific
case!

We could look into moving completely to ES6 or just using let/const in place of the vars, but this
will come with effort, so for now this will do.