use node version from NVM
As per https://github.com/actions/setup-node#v2 'The node-version input is optional. If not supplied, the node version that is PATH will be used.'
So we change here to use nvm to set the node version in the PATH.
This should prevent us needing to duplicate hardcoding the node version.