Update example alternativeFormatContactEmail
The previous example was confused me - I didn't realise it was an example, and thought we were showing the non-existent 'help@' address to real users. Switch the example to match what we're really using in the apps. This should reduce the risk that people in future make the same mistake as I did.